Travis-CI POD Update: 876

+++ JMapiOSSDK4.0/Frameworks/JMapCoreKit.framework/Versions/A/Headers/JMapPoint.h
@@ -8,12 +8,30 @@
 
 #import <Foundation/Foundation.h>
 
+/**
+ *  The JMapPathPerFloor model.
+ */
 @interface JMapPoint : NSObject
-
-@property (nonatomic) double x;
-@property (nonatomic) double y;
-@property (nonatomic) double z;
-
--(id)initWithX:(double)x Y:(double)y Z:(double)z;
+/**
+ *  The x coordinate of the JMapPoint.
+ */
+@property (nonatomic) float x;
+/**
+ *  The y coordinate of the JMapPoint.
+ */
+@property (nonatomic) float y;
+/**
+ *  The map ID of the JMapPoint.
+ */
+@property (nonatomic) float z;
+/**
+ *  Initializes a JMapPoint.
+ *
+ *  @param x The x coordinate of the point.
+ *  @param y The y coordinate of the point.
+ *  @param z The map ID of the point.
+ *  @return An initialized JMapPoint instance.
+ */
+-(id)initWithX:(float)x Y:(float)y Z:(float)z;
 
 @end
